TST: move sparse stack/unstacking tests to sparse/test_reshape
